From Bill:
The attached patch rearranges things so the function
Fl::get_system_colors
parses the switches passed by -fg, -bg, or -bg2.  This allows you to
interpret these switches without using Fl_Window::show(argc,argv), which
is inconvenent for a program that displays many windows, none of which
are the "main" window.