Changes needed to get FLTK to compile on Borland C++ 5 under Windows.  I compiled this on BC++ 5.0 upgraded to 5.0B via the two monster patches. I didn't turn on optimization because my version of BC++ doesn't seem to do much in the way of optimization (FLUID was only 1k smaller when optimized for size).  VC++ generates smaller code.
The examples that use OpenGL don't work because Borland's linker can't find
"wglShareLists".  I'm sure this is a simple problem, but I don't know how
to fix it.

Borland's C++ compiler won't allow you to call main() from C++, so I had
to add a c function in "fl_call_main.c" to call it so that you don't have
to do that WinMain crap.  However, when I added this file to the Visual C++
project it converted the whole thing from 5.0 format to 6.0 format.  The
files look the nearly identical so I don't think this should be a problem
for 5.0 users, but if it is then you can revert them back to the previous
version and just add this one source file.

Borland really doesn't suck that bad.  It doesn't look as polished as VC++
and it refused to supress some warnings for no reason, but I forgot how
much I liked the feel of Borlands compilers...  Much more intuitive then
MS VC++.