New Fl_Color_Chooser with 64 little "cells" at the bottom.  To get a color from a cell, click on it.  To put a color into the cell, click with the middle or right mouse button.  Perhaps a tooltip to explain this?
Fluid uses this new color chooser to allow full 24-bit colors to be
selected for all the widgets.  If you want a gray that is changed by the
-bg switch or other user preferences you must pick one of the color chooser
cells in the third row, these are 0,1/4,1/3,1/2,2/3,3/4,1 indexed grays.
Setting the color chooser to gray wil select a 24-bit gray.

Removed a call to sizes() from the Fl_Group header file that was breaking
the color chooser.  This should not be done, sizes() is designed to be
delayed until the last moment.