Changes by Sylvester Ziolkowski to speed up glx contexts on win32, these changes also seem to speed up Mesa.  It now caches the current context and avoids callig glXMakeCurrent for the same context repeatedly.  Why glX and wgl do not do this internally is a mystery...
This can mess up programs that set the context (for instance by using a
library that draws in OpenGL) because fltk will not realize that the
current context has changed.  I added a call "fl_no_gl_context()" that
can be used to inform fltk that it's cache is no good.

This needs to be tested on WIN32!  If somebody can confirm that this
works I would like to put this into the next release of fltk 1.0 as well.