FL_SYMBOL_LABEL obeys alignment (makes them square too, without the # sign)
Fixed the buttons vanishing in the label demo program (this was a
problem with the forms emulation).

fl_load_style takes the name of a function to call.  This function is
of the form "bool f(const char *)" and should return true if the
plugin is successful.  Fltk calls "fltk_theme" when loading a theme.
Fixed the demonstration themes to use this.