FL_SYMBOL_LABEL obeys alignment (if aligment is set it also makes them square even without the # sign)
Fixed the buttons vanishing in the label demo program (this was a
problem with the forms emulation).

fl_width uses XTextWidth, which should allow kerning fonts to work
(though I suspect X does not support these anyway).  We should use
GetTextExtents on Win32.

Made the sliders & scrollbars highlight when you are pointing at the
trough.

Fixed the pagesize of the scrollbars.

Removed the remaining occurances of \nnn printing of characters from
labels and input widgets.