Changed my mind about that menubar title thing.  What should be happening is that the menubar itself should get shorter or taller based on the font size of the titles.
Of course, in order to do that we need a layout system that does not use
absolute coordinates but resizes and moves widgets based on relationships
with other widgets.  For example, if the menubar gets shorter, the browser
below it should get taller.  Does the new layout code do this?  (I haven't
looked at it yet.)  I know someone wrote a motif-like layout thingie in
the bazaar that does this, perhaps this could be the basis of an easy to
use FLTK version?

I guess what I'm saying that FLTK's default absolute coordinate layout
system ought to be deprecated and replaced with something more modern.
Maybe for FLTK 3.0?

Carl