Removed the highlight value from Fl_Boxtype.  Replaced with a boxdrawing function that draws flat normally and another box type if either highlight or value is on.  The value is passed to this other boxtype, so it can draw up/down as necessary.
Got rid of the cast-away-const stuff in Fl_Menu_Item::draw.  This does
not work anyway as the menu-title flag is not set until the first time
the item is drawn, preventing modifications to the menu title from
inheriting from the menu-title style.  Instead it directly references
the menu-title style for box&color, ignoring the menu item's style.
It also directly references the menu_item default style in some cases
to get around menu items which set the selection color.

Made both the menu items and the menu titles use "selected" colors to
indicate that they are chosen.  This makes them consistent with each
other.  Note that themes that set the highlighted color on menu items
have to change to set the selected color instead.

Made Fl_Style::dynamic() be an inline method.

Moved all the Fl_Style code into it's own source file, also merged in
the Fl_Input_Style code to it.

Added "leading" to the style structure.  This is extra line spacing
between text items.  The Fl_Browser, Fl_Input/Output, and menu windows
use this.  Replaces fl_extra_menu_spacing.  The default is zero except
for menu windows which default to 4.  Unfortunately the menu item box
size is added to this, so if you set the menu item box you probably
need to change this value.  Perhaps this box and the selected color
should be on the menu window style so all the related stuff is in the
same place?