Made fl_rgb(const char*) understand KDE nnn,nnn,nnn format.  This gets rid of the need for a special parser for KDE.  Removed Fl_get_system_colors.cxx and split it into fl_rgb.cxx and fl_background.cxx.  The functions fl_foreground and fl_text_background were removed as they are not used by themes.
I redid most (all?) of the changes Carl accidentally deleted.  In
particular this includes the reservation of FL_KEYUP event number and
the rewriting of the Fl.cxx and Fl_Group.cxx event handling code.

At the moment the wheel moves by 14*3 pixels per 120 units reported by
Windows (or per single click reported by X) to correspond to 3 lines
at standard line spacing.  Still awaiting information about how the UI
really works, as the current one makes very little sense.  I just
deleted the configuration stuff, but if we want it we must move the
"setting" to fl_options.cxx, like all the other settings are.  (I
would prefer not to have UI-behavior options if at all possible, as it
can lead to extreme bloat and make it impossible to debug).

Further investigation of the boxtypes shows that the "down" boxes are also
incompatable with fltk1.0, in that making a "down" button and pushing
it should display an "up" box.  I changed the box definitions to do
this.  This revealed several places in fl_glyph where it was
accidentally sending FL_VALUE to the glyph box, which I suspect is why
Carl had problems with the previous up box.  I tried to fix all the
occurances I found, but there may be a few more.  "highlight_down_box"
will not work (it will go up when pushed) but I could not find any
code that used it, so we may just want to remove it?  I tried the
Alternative theme and scheme in several test programs and did not see
any problems.