Added Fl.H required by an inline function if Fl_Repeat_Button.H
Fl_add_idle adds new functions to the end of the queue ring, rather than the
start, so they are executed in the order added, and a callback that adds
itself does not prevent others from being called.

Fluid lets you type in code that starts with '#' for cpp directives.