EXTENSIVE changes so that the items in menus are normal widgets, which is what everybody really wants.  See the attached email.
As part of the above, the default appearance of the menus was changed
to be more Windows98-like.  In particular the inactive_menu_hack has
been forced to be on because the switching code got very complicated.
You can make individual items not highlight by setting output() on
them, rather than inactive() (this was done for the dividers).

Fl_Choice figures out if motif emulation is on and acts like before in
that case.  Currently it checks to see if box() is FL_UP_BOX for this.

Fl_Window uses it's shortcut() to select the key to close itself.  The
default shortcut is FL_Escape.  Changing it to zero disables the close
on Escape.

Fixed keyboard navigation in the fl_ask and other message popups.

Added two new styles: "Message" and "Message_Icon".  These control the
text and the punctuation icon in the message popups.

Fixed the colors of sliders to be fltk1.0 compatable.

Fixed sevaral calls to Fl_Valuator::step() with arguments like "1/100"
(which is zero, not the .01 I expected).