Fixed the highlighting of overlapping widgets.
However it is still broken if you move the mouse out of the window without
first pointing it at a button.  These problems are leading me to thing that
we may want to disallow overlapping widgets that respond to events (using
them for drawing stuff like in the fluid about box is ok).  The demo program
and perhaps some other old Forms programs would need to be fixed.