Fixed definition of glutBitmapWidth to match header file.
Does not turn visible() on when a window is iconized() or if a modal
window is shown and it's parent is iconized.  This allows the code
"while (w->visible() && w->damage()) Fl::check();" to reliably wait
for the window to be mapped and drawn the first time.

Some comments added to Fl_win32.cxx to explain how the WM_PAINT works.