Updates to Makefiles and shared libraries. * Not only do shared libraries build properly, they also work properly   now! (Mike, can you retest the other Unices again, making sure the SLs   actually work?)
* DLLs are now buildable on Windows in the normal way:

     ./configure --enable-shared --prefix=/Devel
     make
     make install

* DLLs on Windows actually work.

* Themes on Windows actually work (but only when the FLTK libraries are
  built as DLLs and the applications are linked against the DLLs).

Mainly this just involved fixing the headers for the various libraries
with the DLL import & export stuff.  See Fl_Export.H for details.

Had to change Fl::gl_visual() to fl_gl_visual() to avoid sending Fl.H
to DLL hell.  (For an example of hell, see Fl_Shared_Image.H.)  Now in
Fl_Gl_Window.H.

Don't try using theme plugins with statically linked executables,
though.  You've been warned.  (Plugins that don't use FLTK are fine,
though.)

Actually, the method we're using under Unix to build plugins which can be
used by statically linked executables isn't 100% reliable.  Specifically,
I'm talking about compiling parts of the library into ".po" files which
are linked into the plugin.  Essentially what this does is create a
private copy of this code for the theme.  However, if this code has
static variables which much be shared, things may not work properly.  I
think in practice this probably doesn't happen with the type of code we
are linking in, but the potential for problems is there.  However, I left
it as is because I like to be able to use themes in statically linked
FLTK applications...  Windows won't allow this anyway which is why themes
only work correctly with dynamically linked apps.  That is probably the
most "correct" way.

Carl