You can no longer insert ^J,^K,^L,^M into Fl_Int/Float_Input.
Fixed bug noticed by Alexander Rabi Beels, if you are in point-to-type
and move the cursor to a different window with a multiline input and
type up/down, the saved up/down horizontal position from the previous
multiline input was used.  It now resets this on any focus change,
fixing this.

Some code cleanup in Fl_Input.cxx, removed unneeded calls to Fl::compose_reset.