I am not actually making any changes.
This is to inform anybody there that I don't seem to be able to post
email to either the fltk or fltk-dev lists.  But for some reason cvs
is able to post the email.

Hope this can be fixed.

Also I am interested in finding out if anybody is going to Siggraph in
New Orleans?