Added some simple "coding style" options to the Fluid output.  There's a new preferences dialog (Edit/Coding Style) to change how code gets generated. The defaults should be just like it was before, but now you can change where the braces go, whether to use tabs or spaces, how many spaces to use, whether to automatically indent code blocks (like callbacks), and a couple of other minor things.  It's probably not perfect yet, but it's working.
The reason for this is that at work, I have to provide an example for every
major new class or interface that I write.  Since I'm presently working on
FLTK widgets, Fluid is a natural for making the examples.  But the generated
code varies considerably from the company "standard style", which makes it
harder for the other developers to follow.  So by changing these options
a bit, I can get "close enough" to the standard style, right out of Fluid.

This shouldn't add much to the size/speed of Fluid (only about 11K of source
for everything including the prefs dialog), and might make someone's life
easier (I know it does mine.  :-)