Modal and grab stuff *vastly* simplified. See new manual pages. When menus are dismissed by clicking outside of them, events are passed on to the clicked window. Code and behavior on both X and Win32 copied from the Qt source (look for "XGrab" and "XAllow" to see the full horror show!) NYI on windows for clicks in the same program.
I renamed Fl_Window::modal_for() to Fl_Window::child_of().

Fl_Window::set_modal() and Fl_Window::set_non_modal() are emulated
under this new system, but should be avoided.

ComboBox (Fl_Input_Browser) fiddled with a bit to get it to work with
this and for focus to work, but it's not perfect yet...

Removed x.h from Fl_Image.h and fl_draw.h. We don't want this as it
defines many symbols (because of Xlib.h and windows.h header files)
that can cause programs to be non-portable! Required some kludges to
Fl_Image, unfortunately, as the type of some fields are now void*.

Removed Fl_Window::x_root() and y_root(), they do not work and are
unnecessary with the relative positioned widgets.

Made Fl_Widget::label(xxx) not be inline.