Updated Enumerations.h in fltk directory to have some of the mnemonics from the one in the FL directory (why are there two different copies of these things anyway?).
Added a fourth parameter to fl_file_choose() to say whether the dialog produced
is an open dialog or a save dialog.  This is needed on Win32, where the dialog
Windows produces will say "Open" or "Save" depending on how you set a flag.
This parameter just lets fl_file_chooser() know how to set that flag.  The flag
is set to "save" by default, so existing code doesn't have to do anything to
work as it always has.

Added damage() call after expanding a parent item in an Fl_Browser.  This makes
the items below it refresh (they weren't refreshing without it, at least on
Win32).  This doesn't work in release mode, but in debug mode works fine
(even doing a full redraw() doesn't work in release mode -- the
expanding/collapsing item itself redraws, but nothing below it does).

Added header include in fluid.cxx to pull in MAX_PATH definition on Win32.

Added relayout() call after add item in Fl_Menu_.cxx.  Without this, loading
a nice long file in Fluid would open it up with no scroll bars, so you'd have
to expand an item (see other change above) and then use the scroll bar to get
the whole thing to redraw.  There may be a better place for the relayout()
call, but it wasn't obvious to me.