This is intended as a short term fix.
At least on Win32, the latest Fluid hangs every time any of the dialog
windows get reused.  So, for example, add a new function/method, then add a
second function/method, and fluid hangs.

This is a workaround -- it just deletes the dialog windows each time it
finishes with them and recreates them the next time they're needed.  It's
not as efficient, obviously, but Fluid was unusable otherwise.  I haven't
dug into it to find out what the real issue is...