Made fluid compile on NT.
Fl_Guess_Image returns "broken" image when the file does not open or
the data is unrecognized. Remove call to filename_name from it that
should not be there, it now takes literal names only.

Commented out all the calls to cwd() from fluid. This means that all
images have to have absolute path names for now. This must be changed
in the near future but this fixes it for the current file chooser
which always returns absolute path names.

Put fltk::PLASTIC_UP_BOX and fltk::PLASTIC_DOWN_BOX back in.

Added InvisibleBox, which is a more compatable replacement for Fl_Box.

Added ValueOutput, which is much lighter than the old Fl_Value_Output
and is not exclusively for displaying a number in box, the user cannot
change the value any. This is more compatable with old code than
changing the outputs to ValueInput widgets with arrow buttons on them.

Fixes to browser (most visible in fluid where the browser often did
not redraw until you moved the mouse).

Fixed the bug where the focus did not appear when window first
shown. Also fixed the focus navigation in Tabs.  New event
FOCUS_CHANGE. This is sent to groups when the focus is changed to any
widget inside the group.

Made some fixes I needed for the new makefiles. It uses
$(MAKEDEPEND). Turn XDBE on by default (it works on all modern
machines).