Made input buffer pointer to iconv const in Xutf8LookupString emulation to quiet compiler warning about illegal conversion from 'char**' to 'const char**'.
Note: Passing "char" to iconv_open as the source encoding seems to be the
correct way to handle locale dependent encodings from the multibyte buffer.
The iconv spec says it should account for machine dependent endianness and
alignment, and with semantics depending on the OS and the current LC_CTYPE
locale facet.