Added ability to turn off XIM in config.h by setting USE_XIM to 0.
The file xpmFileImage.cxx (?) is missing from laza's checkin.

Not sure about the fill() changes. It would be more interesting to
see a working over() function. If the alpha is stored in the rgb buffer
I would like to indicate it by making alpha pointer == rgb pointer,
as it appears that a quick no-alpha test would be a good idea.

Some indentation fixes, and changes to get it to compile on Unix.