From Matthias:
Added the following files to the archive in order to be able to support
tablets on Windws:

src/pktdef.h src/wintab.h src/wintabx.h

I understand from the text in these files, that they can be freely
distributed as part of fltk:

"The text and information contained in this file may be freely used,
copied, or distributed without compensation or licensing restrictions.
This file is copyright 1991-1998 by LCS/Telegraphics."

Fixed up tablet support on Linux. Added Tablet support on Windows.

New function fltk::event_device() will return
fltk::DEVICE_MOUSE, DEVICE_PEN, or DEVICE_ERASER, depending
on the device that triggered the mouse event.

From spitzak:

4-color version of file chooser. This may be rearranged some more, I
am not real happy with the text fields.

Better tick mark positioning on the sliders.

It crashed if a __windows call was done before open_display. Initialized
the pointers so it does not crash (may call wrong one if no unicode, but
it is probably harmless).