It was not a good idea to make two enumerations called "GRAY". Changed the image one to "LUMINANCE".
drawimage callback must return a pointer to the translated data. This
allows it to return a pointer directly to a buffer, or to uncompress
into the middle of the buffer. Fltk1 compatability could probably be
preserved with a wrapper function for it's calls.

Symbol measurement was making mistakes when the symbol changed the font.
This new one is not perfect but seems to be as good as my previous one
and Leka's changes.

KEYUP is only sent to the focus() widget and any parent of it, same as
KEY events.

SHORTCUT events are sent to the add_event_handler() functions.