Browser can now keep track of a single opened item at each level of the hierarchy, removing the need to have a List store the open flags. You can still store them if you want the user to see more than one open item at a time.
PopupMenu, Choice, and MenuBar could not remember the current item
if it was in a List and not at the top level, and would pick a random
item. Changed so they act like the previous item is the one at the
top level.

Added test/list.cxx program to demonstrate how to write a List to
return your own data and use it in a Browser, PopupMenu, and Choice.