Quick-Fix. ==========
This commit adds fix into fltk-config so it sets proper values in bindir, includedir, libdir and DSOLINK when --prefix 
flag is invoked.
This fix however does not fix --exec-prefix which should check if --prefix is invoked or not. Some developers need 
to set prefix so fltk-config can be used properly. (Usually this is usefull in cross-compiling etc.)
Fix for --exec-prefix is to be expected tomorrow.