last character in CodeEditor (Fluid) would disappear under certain circumstances. The problem is two consecutive calls to Fl_Text_Buffer::insert() which first insert a newline, then indent the text by two spaces. The first 'insert' marks the text all the way to the end for redraw. The second 'insert' adds two characters, however the last character for redraw remains the previously last character, hence the very last two characters are not redrawn.
This fix changes the CodeEditor to do a single 'insert'. It would be better to change Fl_Text_Display to increment the redraw range if more characters are inserted before the actual draw is called. Since this goes too close to core code for my taste, I leave that as an exercise for 2.0.