add_fd() now calls the callback in a parallel thread with the fltk::lock() held on OS/X. This works provided the callback does not call any blocking fltk functions such as fltk::wait().
The previous solution was much too slow and created/killed
a thread on every event and would lock up for 1000 seconds a lot.

There probably is a way to synchronize this so
the select() thread will make an event and then wait for a signal from
the main thread to start select() again, but I don't have time to
figure it out. Another solution is to find some way for OS/X to report
fd-ready as a Carbon event.