Moved the add_leaf and add_group methods from Browser to the base Menu class. I also removed the symbol arguments and added a user_data argument to match the other Menu add() methods. This does not seem to conflict with the use of these by the demo and fluid programs.
The Browser symbols are used at draw time, rather than modifying all
the items in the browser.

I had to restore it using the callback for deselecting items in the
browser, our code relied on this. It is possible that when(WHEN_RELEASE)
will do what you want?

Drag selection and shift and ctrl selection in the multibrowser fixed
for when you click below the last item.