Happy New Year!
New Image object. Changed so that the system allocates the memory
for the pixels and has control over the format. This allows it to
use shared memory on X11 and Windows. You can use the setpixels()
call to write data, or figure out how to write the format (which
will probably be ARGB32) yourself. Makes reading image files into
an Image much easier, and removes 2 intermediate buffers
if shared memory works.

Added "MRGB32" and "RGBM" formats for .png support, but this only
works on OS/X, it is emulated elsewhere by premultiplying the data.

Still needed is the ability to access the pixels and to do things
like Image::inactive(). Also want to clean up SharedImage (there
are 3 ways to add new file types it searches for). And Image is no
longer a drawing surface, so there needs to be a Surface object that
you can draw into and then copy rectangles from.