Changes to make it easier to use images as boxes:
Moved the dx,dy,dw,dh fields to the base Symbol class so you can
change them in an image subclass much more easily.

Removed the focusbox() style, and instead added a new virtual function
called drawOverlay() to symbol to do this. The default version draws
the dotted frame if focus is on.