Initial fixes for window position at or beyond (0,0) under win32.
This is not the whole story, but is better.

We were messing up the initialisation of num_screens on Win32 in that
it was being set to a non-zero value by a call to 
GetSystemMetrics(SM_CMONITORS) but was then subsequently incremented
even higher by each iteration call to the screen_cb(...) function, 
so what we were storing for the screen rectangles was a bit messed up.

Under fltk-1.1, the code used to explicitly reset num_screens to zero
before starting to iterate the screen_cb(...) function, so that worked OK.

I have sort-of restored that behaviour, and removed the (now redundant)
call to GetSystemMetrics(SM_CMONITORS) since we *always* enumerate the 
monitors now to get the per-screen DPI values out.

I've also put in a commented out stub where we could store the per-screen 
work-area, which it seems we do also need now, though I have not yet 
contrived to actually implement that.
Volunteers welcomed...