Removes a constraint that the fancy string drawing function fl_draw() is limited to 1024 chars/line that blocked a user of the Fl_Browser widget (see "fl_draw MAXBUF limit" in fltk.coredev).
Also, removed a useless computation in string expansion that checked for valid UTF-8 sequences:
the point is that a valid UTF-8 sequence for a non-ascii char contains no ascii char,
thus no tab, space, control, & or @ we want to process differently. 
Also, invalid UTF-8 sequences are copied unchanged by this procedure.
Therefore, checking for tab, space, control, & or @, and copying the byte otherwise, is enough.