Adjust the Fl_Preferences method
char Fl_Preferences::RootNode::getPath( char *path, int pathlen )

so that it returns a (1), rather than a (-1) in the case that
RUNTIME preferences are being read.

This is for several reasons:

- On some platforms, char is unsigned, so a method that is spec'd to 
return a char type can not safely return (-1) as a value.

- It seems like (most?) other fltk methods that return 
char true/false results return 0 or 1, so this is more consistent.


Thanks to Michael Baeuerle for spotting this one.