Fix documentation of Fl_Input_::word_start() and word_end().
Documentation changes in svn r10383 (STR #3014) said that this would skip
"space-separated" words, which is not true.

Note: The set of word separators is currently not documented, it is
inconsistent with Fl_Text_Display/Fl_Text_Editor, and not UTF-8-aware.
Also, word selection and the above methods use different word delimiters.
This should be fixed in a later release (after 1.3.3).