Add linker option -Wl,-no-undefined if supported (with configure check).
Experimental. Thanks to Greg for finding this option.

If this option is supported by the linker, it is used when linking shared
libraries like libfltk_gl.so. With this option linking of the shared library
fails instead of postponing failure to linking programs with this library,
if any undefined references exist in this library. See STR #3191.