porting branch: Modify unittest_images.cxx to support "on the fly" update of the image test controls.
This builds on the mechanism that Albrecht added to test flipped and inverted image drawing, but allows the settings to be changed at runtime rather than at compile time as previously.

I'm commiting it here for testing use in the porting branch, but I'm not happy I've really tested it enough yet so reluctant to commit it in the main fltk-1.3.x branch until this code has been reviewed and tested by others.